requirements.txt
numpy==2.1.1
deap==1.4.1
gymnasium==1.0.0a2
matplotlib==3.9.2
networkx==3.2.1
pulp==2.7.0
scipy==1.14.1
requests==2.32.3
ell-ai==0.0.14
openai==1.51.0